# [Proof of Concept] Using PointNeighbors.jl instead of a vector of bonds