--- a/src/physics/bond_based.jl
+++ b/src/physics/bond_based.jl
@@ -126,3 +126,35 @@ function force_density_point!(storage::BBVerletStorage, system::BondSystem, ::BB
     end
     return nothing
 end
+
+function force_density_point_inner!(storage::BBVerletStorage, system::BondSystem, ::BBMaterial,
+                              params::BBPointParameters, i::Int)
+    for bond_id in each_bond_idx(system, i)
+        bond = system.bonds[bond_id]
+        j, L = bond.neighbor, bond.length
+
+        # current bond length
+        Δxijx = storage.position[1, j] - storage.position[1, i]
+        Δxijy = storage.position[2, j] - storage.position[2, i]
+        Δxijz = storage.position[3, j] - storage.position[3, i]
+        l = sqrt(Δxijx * Δxijx + Δxijy * Δxijy + Δxijz * Δxijz)
+
+        # bond strain
+        ε = (l - L) / L
+
+        # failure mechanism
+        # if ε > params.εc && bond.fail_permit
+        #     storage.bond_active[bond_id] = false
+        # end
+        # storage.n_active_bonds[i] += storage.bond_active[bond_id]
+
+        # update of force density
+        scfactor = 1
+        bond_fail = true
+        temp = bond_fail * scfactor * params.bc * ε / l * system.volume[j]
+        storage.b_int[1, i] += temp * Δxijx
+        storage.b_int[2, i] += temp * Δxijy
+        storage.b_int[3, i] += temp * Δxijz
+    end
+    return nothing
+end